Starting to improve API
