+++ velhia.api/src/config/routes/algorithmRoutes.ts
@@ -5,9 +5,7 @@ import { AlgorithmController } from '@useCases/algorithm/algorithmController'
 const algorithmRoutes = Router()
 const algorithmController = container.resolve(AlgorithmController)
 
-algorithmRoutes.get('/', (req, res) => algorithmController.getAll(req, res))
-algorithmRoutes.get('/:id', (req, res) => algorithmController.getOne(req, res))
-algorithmRoutes.get('/limit/:limit', (req, res) => algorithmController.getLast(req, res))
+algorithmRoutes.get('/', (req, res) => algorithmController.get(req, res))
 algorithmRoutes.post('/', (req, res) => algorithmController.create(req, res))
 algorithmRoutes.put('/:id', (req, res) => algorithmController.update(req, res))
 algorithmRoutes.delete('/:id', (req, res) => algorithmController.delete(req, res))
