+++ velhia.api/src/config/routes/matchRoutes.ts
@@ -5,9 +5,7 @@ import { MatchController } from '@useCases/match/matchController'
 const matchRoutes = Router()
 const matchController = container.resolve(MatchController)
 
-matchRoutes.get('/', (req, res) => matchController.getAll(req, res))
-matchRoutes.get('/:id', (req, res) => matchController.getOne(req, res))
-matchRoutes.get('/limit/:limit', (req, res) => matchController.getLast(req, res))
+matchRoutes.get('/', (req, res) => matchController.get(req, res))
 matchRoutes.post('/', (req, res) => matchController.create(req, res))
 matchRoutes.put('/:id', (req, res) => matchController.update(req, res))
 matchRoutes.delete('/:id', (req, res) => matchController.delete(req, res))
