--- velhia.api/src/domain/interfaces/iMAS.ts
@@ -1,43 +1,0 @@
-interface IMAS {
-  family: {
-    institution: string,
-    id: string,
-    generation: string,
-    progenitor: string,
-    life: number,
-    memories: number,
-    matchAsLearner: number,
-    matchAsLeader: number,
-    victories: number,
-    defeats: number,
-    draws: number
-  },
-  education: {
-    institution: string,
-    id: string,
-    generation: string,
-    progenitor: string,
-    life: number,
-    memories: number,
-    matchAsLearner: number,
-    matchAsLeader: number,
-    victories: number,
-    defeats: number,
-    draws: number
-  },
-  religion: {
-    institution: string,
-    id: string,
-    generation: string,
-    progenitor: string,
-    life: number,
-    memories: number,
-    matchAsLearner: number,
-    matchAsLeader: number,
-    victories: number,
-    defeats: number,
-    draws: number
-  }
-}
-
-export default IMAS
