+++ velhia.api/src/useCases/education/educationController.ts
@@ -13,57 +13,15 @@ export class EducationController {
 
   /**
    * Receive a get request and
-   * return all Agent saved in the db
+   * return all Agent filter by a query in the db
    * @param {Request} req request
    * @param {Response} res response
-   * @example getAll()
+   * @example get()
    * @returns {Promise<Response | void>} IAgent[] | 404 
    */
-  async getAll(req: Request, res: Response): Promise<Response | void> {
+  async get(req: Request, res: Response): Promise<Response | void> {
     try {
       const mas: IAgent[] | null = await this.repository.getAllAgent()
-      if (mas == null) {
-        return res.sendStatus(404)
-      }
-      return res.send(mas)
-    } catch (error) {
-      res.sendStatus(400).send(error)
-    }
-  }
-
-  /**
-   * Receive a get request with id and
-   * return this Agent saved in the db
-   * @param {Request} req request
-   * @param {Response} res response
-   * @example getOne('5f9db3a3fc7c860a3e316712')
-   * @returns {Promise<Response | void>} IAgent | 404
-   */
-  async getOne(req: Request, res: Response): Promise<Response | void> {
-    try {
-      const masId: string = req.params.id
-      const mas: IAgent | null = await this.repository.getOneAgent(masId)
-      if (mas == null) {
-        return res.sendStatus(404)
-      }
-      return res.send(mas)
-    } catch (error) {
-      res.sendStatus(400).send(error)
-    }
-  }
-
-  /**
-   * receive a get request with limit X to get the  
-   * X latest Agent created in the db
-   * @param {Request} req request
-   * @param {Response} res response
-   * @example getLast(2)
-   * @returns {Promise<Response | void>} IAgent[2] | 404
-   */
-  async getLast(req: Request, res: Response): Promise<Response | void> {
-    try {
-      const limit: number = +req.params.limit
-      const mas: IAgent[] | null = await this.repository.getLastAgent(limit)
       if (mas == null) {
         return res.sendStatus(404)
       }
