+++ velhia.api/src/useCases/match/matchController.ts
@@ -13,57 +13,15 @@ export class MatchController {
 
   /**
    * Receive a get request and
-   * return all Match saved in the db
+   * return all Match filter by a query in the db
    * @param {Request} req request
    * @param {Response} res response
-   * @example getAll()
+   * @example get()
    * @returns {Promise<Response | void>} IMatch[] | 404 
    */
-  async getAll(req: Request, res: Response): Promise<Response | void> {
+  async get(req: Request, res: Response): Promise<Response | void> {
     try {
       const mac: IMatch[] | null = await this.repository.getAllMatch()
-      if (mac == null) {
-        return res.sendStatus(404)
-      }
-      return res.send(mac)
-    } catch (error) {
-      res.sendStatus(400).send(error)
-    }
-  }
-
-  /**
-   * Receive a get request with id and
-   * return this Match saved in the db
-   * @param {Request} req request
-   * @param {Response} res response
-   * @example getOne('5f9db3a3fc7c860a3e316712')
-   * @returns {Promise<Response | void>} IMatch | 404
-   */
-  async getOne(req: Request, res: Response): Promise<Response | void> {
-    try {
-      const macId: string = req.params.id
-      const mac: IMatch | null = await this.repository.getOneMatch(macId)
-      if (mac == null) {
-        return res.sendStatus(404)
-      }
-      return res.send(mac)
-    } catch (error) {
-      res.sendStatus(400).send(error)
-    }
-  }
-
-  /**
-   * receive a get request with limit X to get the  
-   * X latest Match created in the db
-   * @param {Request} req request
-   * @param {Response} res response
-   * @example getLast(2)
-   * @returns {Promise<Response | void>} IMatch[2] | 404
-   */
-  async getLast(req: Request, res: Response): Promise<Response | void> {
-    try {
-      const limit: number = +req.params.limit
-      const mac: IMatch[] | null = await this.repository.getLastMatch(limit)
       if (mac == null) {
         return res.sendStatus(404)
       }
